sif-chatbot/requirements.txt
gradio>=3.50.2
pandas>=2.0.0
numpy>=1.24.0
spacy>=3.7.2
scikit-learn>=1.3.0
python-dotenv>=1.0.0
office365-rest-python-client>=2.4.0
schedule>=1.2.0
sqlalchemy>=2.0.0